% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfComputation.R
\name{ccfComparison}
\alias{ccfComparison}
\title{Computes the location of the maximum CCF value in a CCF map between two
matrices}
\usage{
ccfComparison(mat1, mat2, ccfMethod = "fft")
}
\arguments{
\item{ccfMethod}{implements 3 different methods to calculate the CCF -- all which yield differing CCF values}
}
\description{
Computes the location of the maximum CCF value in a CCF map between two
matrices
}
\seealso{
cartridges3D package \url{https://github.com/xhtai/cartridges3D}
}
\keyword{internal}
