% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{cellDivision}
\alias{cellDivision}
\title{Divides an image into cells}
\usage{
cellDivision(surfaceMat, horizSplits = 8, vertSplits = 8, ...)
}
\arguments{
\item{surfaceMat}{surface matrix of a breech face impression}

\item{horizSplits}{number of splits along horizontal axis}

\item{vertSplits}{number of splits along vertical axis}
}
\description{
This is a helper for the cellCCF function.
}
