% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{checkForBreechface}
\alias{checkForBreechface}
\title{Determines whether a matrix contains enough non-NA values.}
\usage{
checkForBreechface(cell, minObservedProp = 0.15)
}
\arguments{
\item{cell}{a matrix representing a cell from a breech face impression image}
}
\description{
This is a helper for the cellCCF function. This function
returns TRUE if a given cell contains more than 225 pixels of observed
values (i.e., non-NA values) and FALSE otherwise.
}
\keyword{internal}
