% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfComputation.R
\name{circshift}
\alias{circshift}
\title{Performs a circular shift based on periodic boundary conditions in a vector}
\usage{
circshift(x, vec)
}
\description{
Performs a circular shift based on periodic boundary conditions in a vector
}
\seealso{
http://stackoverflow.com/questions/18791212/equivalent-to-numpy-roll-in-r
}
\keyword{internal}
