% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{cmcFilter}
\alias{cmcFilter}
\title{Applies Congruent Matching Cells method logic to the CCF results of a
comparison between two cartridge case scans.}
\usage{
cmcFilter(
  ccfResults,
  consensus_function = median,
  ccf_thresh = 0.4,
  dx_thresh = 20,
  dy_thresh = dx_thresh,
  theta_thresh = 3,
  consensus_function_theta = consensus_function,
  ...
)
}
\arguments{
\item{...}{arguments to be passed to consensus_function or
consensus_function_theta if necessary}
}
\description{
Applies Congruent Matching Cells method logic to the CCF results of a
comparison between two cartridge case scans.
}
