% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{cmcFilter_improved}
\alias{cmcFilter_improved}
\title{Implements "improved" CMC logic on a list of CCF results for a comparison
between two cartridge case scans as proposed by Tong et al. (2015)}
\arguments{
\item{consensus_function}{function to aggregate the translation (x and y)
and rotation (theta) values in the ccfDF data frame to determine
"consensus" values}

\item{ccf_thresh}{minimum correlation threshold to call a cell pair
"congruent matching"}

\item{dx_thresh}{maximum distance from the consensus x value that a cell
pair can be to be called "congruent matching"}

\item{dy_thresh}{maximum distance from the consensus y value that a cell
pair can be to be called "congruent matching"}

\item{thetaThresh}{maximum distance from the consensus theta value that a
cell pair can be to be called "congruent matching"}

\item{missingThetaDecision}{dictates how function should handle situations
in which one direction passes the high CMC criterion while another
direction does not. "replace": replaces theta value in failed direction
with opposite of theta value in successful direction. "dismiss": only
counts the initial CMCs in failed direction and high CMCs in successful
direction. "fail": only counts the initial CMCs in either direction.}

\item{compareThetas}{dictates if the consensus theta values
determined under the initially proposed method should be compared to the
consensus theta values determined under the High CMC method. In particular,
determines for each direction whether the consensus theta values determined
under the two methods are within thetaThresh of each other. It is often
the case that non-matching cartridge cases, even if they pass the High CMC
criterion, will have differing consensus theta values under the two
methods. If this isn't taken into account, non-matches tend to be assigned
a lot of false positive CMCs under the High CMC method.}

\item{consensus_function_theta}{\emph{(OPTIONAL)} function (separate from
consensus_function) to aggregate the rotation (theta) values in the ccfDF
data frame to determine "consensus" value}
}
\description{
Implements "improved' Congruent Matching Cells logic, as
proposed by Tong et al. (2015), to the CCF results of a comparison between
two cartridge case scans.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4730689/pdf/jres.120.008.pdf}
}
\keyword{internal}
