% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{cmcPerThetaBarPlot}
\alias{cmcPerThetaBarPlot}
\title{Create a bar plot of congruent matching cells per rotation value}
\usage{
cmcPerThetaBarPlot(
  cellCCF_output,
  consensus_function = median,
  corr_thresh = 0.5,
  dx_thresh = 15,
  dy_thresh = dx_thresh,
  theta_thresh = 3,
  consensus_function_theta = consensus_function,
  highCMCThresh = 1
)
}
\arguments{
\item{cellCCF_output}{list returned by the cellCCF or cellCCF_bothDirections
function. If from the cellCCF_bothDirections, then the ggplot will be
faceted by the "direction" of the comparison (i.e., whether x3p1 or x3p2
played the role as the "questioned" cartridge case scan)}
}
\description{
Create a bar plot of congruent matching cells per rotation value
}
\examples{
\dontrun{
 #x3p1 and x3p2 are two x3p objects containing processed cartridge case scans
 comparison1 <- cellCCF(x3p1,x3p2)

 #creates a single bar plot of the CMCs per theta
 cmcPerThetaBarPlot(comparison1)

 comparison2 <- cellCCF_bothDirections(x3p1,x3p2)

 #creates a faceted bar plot of the CMCs per theta in both comparison directions
 cmcPerThetaBarPlot(comparison2)
}

}
