% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfComputation.R
\name{comparison}
\alias{comparison}
\title{Computes the location of the maximum CCF value in a CCF map between two
matrices}
\usage{
comparison(im1, im2)
}
\description{
Computes the location of the maximum CCF value in a CCF map between two
matrices
}
\keyword{internal}
