% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_fft.ccf}
\alias{comparison_fft.ccf}
\title{Estimate translation alignment between a cell/region pair based on the
\href{https://mathworld.wolfram.com/Cross-CorrelationTheorem.html}{Cross-Correlation Theorem}}
\usage{
comparison_fft.ccf(cellHeightValues, regionHeightValues)
}
\description{
Estimate translation alignment between a cell/region pair based on the
\href{https://mathworld.wolfram.com/Cross-CorrelationTheorem.html}{Cross-Correlation Theorem}
}
\note{
The FFT is not defined for matrices containing missing values. The
missing values in the cell and region need to be replaced before using this
function. See the \link\href{comparison_standardizeHeightValues}{cmcR} function
to replace missing values after standardization.
}
