% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_replacingMissingValues}
\alias{comparison_replacingMissingValues}
\title{Replace missing values in a scan}
\usage{
comparison_replacingMissingValues(heightValues, replacement = 0)
}
\description{
Replace missing values in a scan
}
