% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_standardizeHeight}
\alias{comparison_standardizeHeight}
\alias{comparison_standardizeHeights}
\title{Standardize height values of a scan by centering/scaling by desired
statistics and replacing missing values}
\usage{
comparison_standardizeHeights(
  heightValues,
  withRespectTo = "individualCell",
  centerBy = mean,
  scaleBy = sd
)
}
\description{
Standardize height values of a scan by centering/scaling by desired
statistics and replacing missing values
}
\note{
this function adds information to the metainformation of the x3p scan
it is given that is required for calculating, for example, the
pairwise-complete correlation using the comparison_cor function.
}
