% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_standardizeHeightValues}
\alias{comparison_standardizeHeightValues}
\title{Standardize height values of a scan by centering/scaling by desired
statistics and replacing missing values}
\usage{
comparison_standardizeHeightValues(
  heightValues,
  withRespectTo = "individualCell",
  centerBy = mean,
  scaleBy = sd,
  replaceMissingValues = TRUE
)
}
\description{
Standardize height values of a scan by centering/scaling by desired
statistics and replacing missing values
}
