% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{decision_CMC}
\alias{decision_CMC}
\title{Applies the decision rules of the original method of Song (2013) or the High
CMC method of Tong et al. (2015)}
\usage{
decision_CMC(
  cellIndex,
  x,
  y,
  theta,
  corr,
  xThresh = 20,
  yThresh = xThresh,
  thetaThresh = 6,
  corrThresh = 0.5,
  tau = NULL
)
}
\arguments{
\item{tau}{(optional) parameter required to apply the High CMC method of Tong
et al. (2015). If not given, then the decision rule of the original method
of Song (2013) is applied.}
}
\description{
Applies the decision rules of the original method of Song (2013) or the High
CMC method of Tong et al. (2015)
}
