% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{decision_combineCMCDirections}
\alias{decision_combineCMCDirections}
\title{Combine CMC data from two comparison directions}
\usage{
decision_combineCMCDirections(
  comparison_1to2_df,
  comparison_2to1_df,
  corColName = "pairwiseCompCor",
  method = "highCMC",
  compareThetas = TRUE,
  thetaThresh = 6
)
}
\description{
Combine CMC data from two comparison directions
}
