% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{decision_highCMC_identifyHighCMCThetas}
\alias{decision_highCMC_identifyHighCMCThetas}
\title{Classify theta values in CMC-theta distribution as having "High" or "Low" CMC
candidate counts}
\usage{
decision_highCMC_identifyHighCMCThetas(cmcThetaDistrib, tau = 1)
}
\description{
Classify theta values in CMC-theta distribution as having "High" or "Low" CMC
candidate counts
}
