% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{decision_originalMethod}
\alias{decision_originalMethod}
\alias{decision_originalMethod_classifyCMCs}
\title{Apply CMC classification logic of the original method of Song (2013)}
\usage{
decision_originalMethod_classifyCMCs(
  cellIndex,
  x,
  y,
  theta,
  corr,
  xThresh = 20,
  yThresh = xThresh,
  thetaThresh = 6,
  corrThresh = 0.5
)
}
\description{
Apply CMC classification logic of the original method of Song (2013)
}
\seealso{
\url{https://tsapps.nist.gov/publication/get_pdf.cfm?pub_id=911193}
}
