% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision_originalMethod_classifyCMCs}
\alias{decision_originalMethod_classifyCMCs}
\title{Apply CMC classification logic of the original method of Song (2013)}
\usage{
decision_originalMethod_classifyCMCs(
  cellIndex,
  x,
  y,
  theta,
  corr,
  xThresh = 20,
  yThresh = xThresh,
  thetaThresh = 6,
  corrThresh = 0.5
)
}
\arguments{
\item{cellIndex}{vector/tibble column containing cell indices corresponding
to a reference cell}

\item{x}{vector/tibble column containing x horizontal translation values}

\item{y}{vector/tibble column containing y vertical translation values}

\item{theta}{vector/tibble column containing theta rotation values}

\item{corr}{vector/tibble column containing correlation similarity scores
between a reference cell and its associated target region}

\item{xThresh}{used to classify particular x values "congruent" if they are
within xThresh of the median x value}

\item{yThresh}{used to classify particular y values "congruent" if they are
within yThresh of the median y value}

\item{thetaThresh}{used to classify particular theta values "congruent" if
they are within thetaThresh of the median theta value}

\item{corrThresh}{to classify particular correlation values "congruent" if
they are at least corrThresh}
}
\description{
Apply CMC classification logic of the original method of Song (2013)
}
\note{
the decision_CMC function internally calls this function if a value of
tau is not provided
}
\examples{
data(fadul1.1_processed,fadul1.2_processed)

comparisonDF <- purrr::map_dfr(seq(-30,30,by = 3),
                              ~ comparison_allTogether(fadul1.1_processed,
                                                       fadul1.2_processed,
                                                       theta = .))


comparisonDF <- comparisonDF \%>\%
dplyr::mutate(originalMethodClassif = decision_originalMethod_classifyCMCs(cellIndex = cellIndex,
                                                                           x = x,
                                                                           y = y,
                                                                           theta = theta,
                                                                           corr = pairwiseCompCor))

comparisonDF \%>\%
dplyr::filter(originalMethodClassif == "CMC")
}
\seealso{
\url{https://tsapps.nist.gov/publication/get_pdf.cfm?pub_id=911193}
}
