% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{extractCellbyCornerLocs}
\alias{extractCellbyCornerLocs}
\title{Cuts out a cell in a surface matrix.}
\usage{
extractCellbyCornerLocs(cornerLocs, rotatedSurfaceMat, mat2Dim)
}
\arguments{
\item{cornerLocs}{the location (indices) of the desired cell in the image}

\item{rotatedSurfaceMat}{a surface matrix representing a rotated breech face
impression image}
}
\description{
This is a helper for the cellCCF function. This function
cuts out a cell in a surface matrix based on the most top, bottom, left,
and right indices of the cell's location in the surface matrix.
}
\keyword{internal}
