% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fadulData_raw}
\alias{fadulData_raw}
\alias{fadul1.1_raw}
\alias{fadul1.2_raw}
\alias{fadul2.1_raw}
\title{"Fadul 1-1", "Fadul 1-2", and "Fadul 2-1" cartridge cases from Fadul et al.
(2011) in their raw format}
\format{
An x3p object containing a surface matrix and metainformation
concerning the conditions under which the scan was taken \describe{
\item{header.info}{size and resolution of scan}

\item{surface.matrix}{spatially-ordered matrix of elements representing the
height values of the processed cartridge case surface at particular
locations}

\item{feature.info}{provides structure for storing surface data}

\item{general.info}{information concerning the author of the scan and
capturing device}

\item{matrix.info}{provides link to surface measurements in binary format}

}

An object of class \code{x3p} of length 5.

An object of class \code{x3p} of length 5.
}
\source{
\url{https://tsapps.nist.gov/NRBTD/Studies/CartridgeMeasurement/Details/2d9cc51f-6f66-40a0-973a-a9292dbee36d}
}
\usage{
fadul1.1_raw

fadul1.2_raw

fadul2.1_raw
}
\description{
An x3p (XML 3D Surface Profile) object containing the "Fadul 1-1", "Fadul
1-2", and "Fadul 2-1" cartridge cases from Fadul et al. (2011) in their raw
format
}
\seealso{
T. Fadul, G. Hernandez, S. Stoiloff, and G. Sneh. An Empirical Study
to Improve the Scientific Foundation of Forensic Firearm and Tool Mark
Identification Utilizing 10 Consecutively Manufactured Slides, 2011.

\url{https://github.com/heike/x3ptools}
}
\keyword{datasets}
