% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfComputation.R
\name{filterViaFFT}
\alias{filterViaFFT}
\title{Calculates cross-correlation between two matrices using FFTs}
\usage{
filterViaFFT(A, B)
}
\description{
Calculates cross-correlation between two matrices using FFTs
}
\keyword{internal}
