% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_info.R
\name{get_exchange_info}
\alias{get_exchange_info}
\title{Retrieve the information about a given exchange}
\usage{
get_exchange_info(exchange_id = 2, api_key = Sys.getenv("CMC_API_KEY"))
}
\arguments{
\item{exchange_id}{The exchange ID as returned by the get_exchange_map function}

\item{api_key}{The api key. If not provided it will check if it is an
environmental variable.}
}
\value{
A row tibble contianing the information. This allows for the use of
map_dfr function to iterate over many exchanges
}
\description{
this function extracts the description, start data, number of weekly visits
and what fiats it accepts. The last two are at the point in time when the
function is called and are not historic.
}
\examples{
\dontrun{
  x <- get_exchange_info(exchange_id = 2)
}
}
