% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exchange_map.R
\name{get_exchange_map}
\alias{get_exchange_map}
\title{Return the map of all exchanges in CMC}
\usage{
get_exchange_map(api_key = Sys.getenv("CMC_API_KEY"))
}
\arguments{
\item{api_key}{The API key. If not provided it looks in environmental variables}
}
\value{
Returns a tibble with the information
}
\description{
This function returns the CMC ID Map for all the currencies. Specifically
it returns the metadata related to them (activity when called, start date
 end data and platform). Returns both active and inactive coins
}
\examples{
\dontrun{
  x <- get_exchange_map()
}
}
