% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_metrics.R
\name{get_global_metrics}
\alias{get_global_metrics}
\title{Get the global metrics for the universe of assets}
\usage{
get_global_metrics(
  time_start = NULL,
  time_end = NULL,
  interval = "daily",
  api_key = Sys.getenv("CMC_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{time_start}{The start of the interval to grab (defaults to today)}

\item{time_end}{The end of the interval to grab (defaults to today)}

\item{interval}{The frequency to grab (defaults to daily)}

\item{api_key}{The API key. If not provided it looks in environmental variables}

\item{as_date}{logical. Convert returned time stamps to dates?}
}
\value{
Returns a tibble with the information
}
\description{
Get the global metrics for the universe of assets
}
\examples{
\dontrun{
  x <- get_global_metrics()
}
}
