% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_metrics.R
\name{get_key_metrics}
\alias{get_key_metrics}
\title{Get total credits available and remaining credits}
\usage{
get_key_metrics(api_key = Sys.getenv("CMC_API_KEY"))
}
\arguments{
\item{api_key}{character. CMC Pro API key. Defaults to the
\code{CMC_API_KEY} environment variable.}
}
\value{
a tibble that contains credits remain and amount used on a per minute and monthly basis.
}
\description{
Get total credits available and remaining credits
}
\examples{
\dontrun{
  #Requires a valid API key to run
  x <- get_key_metrics()
}
}
