% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token_quote.R
\name{get_token_quote}
\alias{get_token_quote}
\title{This function gets the price and volume information of a given token}
\usage{
get_token_quote(
  token = "BTC",
  time_start = NULL,
  time_end = NULL,
  interval = "daily",
  api_key = Sys.getenv("CMC_API_KEY"),
  as_date = TRUE
)
}
\arguments{
\item{token}{The official token symbol from CMC for the token}

\item{time_start}{What is the earliest date to retrieve}

\item{time_end}{What is the latest date to retrieve}

\item{interval}{The interval to retrieve}

\item{api_key}{The API key to use}

\item{as_date}{Should dates be returned instead ot a POSIXct for daily series}
}
\value{
A tibble containing the data
}
\description{
This function takes the name of _one_ token and returns a time series
tibble of the price, 24h volume, marketcap, and supply of the token. The
data is at a daily level
}
\details{
For the moment it only takes one token. But hopefully will fix this in
the next release
}
\examples{
\dontrun{
  x <- get_token_quote("BTC", time_start = "2024-01-01", time_end = "2024-01-31")
}
}
