% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{getCellRegionPairs}
\alias{getCellRegionPairs}
\title{Extract cell/region pair matrices for a comparison of interest}
\arguments{
\item{x3p1}{an x3p object}

\item{x3p2}{another x3p object}

\item{ccfDF}{data frame containing comparison results between x3p1 and x3p2. At a
minimum, this needs to contain "cellNum" and "theta" columns.}

\item{cellCCF_params}{list of parameters under which the comparison represented in
ccfDF was performed. Such a list is returned by the cellCCF and
cellCCF_bothDirections function}
}
\description{
This function is meant as a diagnostic tool to determine what the
cell/region pairs for a particular cartridge case comparison looked like
right before calculation of the CCF.
}
\examples{
\dontrun{
comparison <- cellCCF(x3p1,x3p2)

topResults <- comparison$ccfResults \%>\%
 cmcR::topResultsPerCell()

cellRegionPairs <- getCellRegionPairs(x3p1,x3p2,topResults,comparison$params)
}

}
