% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{getMode}
\alias{getMode}
\title{Calculates the mode of a vector of numbers}
\usage{
getMode(x)
}
\description{
Calculates the mode of a vector. Can be used as a consensus
function in cmcR::cmcFilter or cmcR::cmcFilter_improved.
}
