% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{levelBFImpression}
\alias{levelBFImpression}
\title{Helper for selectBFImpression}
\usage{
levelBFImpression(ransacFit, useResiduals = TRUE, ...)
}
\arguments{
\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}
}
\description{
Helper for selectBFImpression
}
\keyword{internal}
