% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firingPinCircleDetection.R
\name{nms}
\alias{nms}
\title{Non-maxima suppression}
\usage{
nms(im, sigma)
}
\description{
Non-maxima suppression
}
\seealso{
#https://www.rdocumentation.org/packages/imager/versions/0.41.2/topics/hough_circle
}
\keyword{internal}
