% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfComputation.R
\name{pairwise_ccf}
\alias{pairwise_ccf}
\title{calculates the "brute force" pairwise-complete correlation between matrix a
and matrix b translated by (offsetx,offsety)}
\usage{
pairwise_ccf(a, b, offsetx = 0, offsety = 0)
}
\description{
calculates the "brute force" pairwise-complete correlation between matrix a
and matrix b translated by (offsetx,offsety)
}
\keyword{internal}
