% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSegmentation.R
\name{preProcess_cropBFExterior}
\alias{preProcess_cropBFExterior}
\title{Crop the exterior of a breech face impression surface matrix}
\usage{
preProcess_cropBFExterior(
  x3p,
  scheme = 3,
  high_connectivity = FALSE,
  tolerance = 0,
  radiusOffset = 0,
  croppingThresh = 1,
  agg_function = median
)
}
\arguments{
\item{x3p}{an x3p object containing the surface matrix of a cartridge case
scan}

\item{scheme}{argument for imager::imgradient}

\item{high_connectivity}{argument for imager::label}

\item{tolerance}{argument for imager::label}

\item{radiusOffset}{number of pixels to add to estimated breech face radius.
This is commonly a negative value (e.g., -20) to trim the cartridge case
primer roll-off from the returned, cropped surface matrix.}

\item{croppingThresh}{argument for cmcR::preProcess_cropWS}

\item{agg_function}{the breech face radius estimation procedure returns
numerous radius estimates. This argument dictates the function used to
aggregate these into a final estimate.}
}
\value{
An x3p object containing the surface matrix of a breech face
impression scan where the rows/columns on the exterior of the breech face
impression have been cropped-out.
}
\description{
The radius estimation procedure tends to over-estimate the
desired radius values. As such, a lot of the breech face impression
"roll-off" is included in the final scan. Excessive roll-off can bias the
calculation of the CCF. As such, we can manually shrink the radius estimate
so that little to no roll-off is included in the final processed scan.
}
\examples{
fadul1.1 <- x3ptools::read_x3p("https://tsapps.nist.gov/NRBTD/Studies/CartridgeMeasurement/DownloadMeasurement/2d9cc51f-6f66-40a0-973a-a9292dbee36d")

fadul1.1_cropped <- preProcess_cropBFExterior(x3p = fadul1.1,radiusOffset = -20)

x3pListPlot(list("Original" = fadul1.1,"Cropped" = fadul1.1_cropped))
}
