% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_cropWS}
\alias{preProcess_cropWS}
\title{Crop out rows/columns outside of the breech face impression in a cartridge
case scan.}
\usage{
preProcess_cropWS(surfaceMat, croppingThresh = 2, ...)
}
\arguments{
\item{surfaceMat}{a surface matrix representing a breech face impression scan}

\item{croppingThresh}{minimum number of non-NA pixels that need to be in a
row/column for it to not be cropped out of the surface matrix}
}
\description{
Crop out rows/columns outside of the breech face impression in a cartridge
case scan.
}
