% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess_erode}
\alias{preProcess_erode}
\title{Erode the interior or exterior of a cartridge case surface}
\usage{
preProcess_erode(x3p, region, morphRadius = 50)
}
\arguments{
\item{x3p}{an x3p object}

\item{region}{either "interior," meaning the observations around the firing
pin hole will be eroded, or "exterior," meaning the observations around the
outer edge of the cartridge case primer will be eroded}

\item{morphRadius}{controls the amount of erosion. Larger values correspond to
a larger (circular) morphological mask leading to more erosion.}
}
\description{
performs the morphological operations and dilation to "shave"
observations off of the interior or exterior of a cartridge case surface
matrix.
}
