% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess_filterInterior}
\alias{preProcess_filterInterior}
\title{Filter-out the firing pin impression region of a breech face impression scan}
\usage{
preProcess_filterInterior(
  x3p,
  scheme = 3,
  high_connectivity = FALSE,
  tolerance = 0,
  radiusOffset = 0
)
}
\arguments{
\item{x3p}{an x3p object containing the surface matrix of a cartridge case
scan}

\item{scheme}{argument for imager::imgradient}

\item{high_connectivity}{argument for imager::label}

\item{tolerance}{argument for imager::label}

\item{radiusOffset}{number of pixels to add to estimated firing pin hole
radius. This is commonly a positive value (e.g., 200) to not only remove
observations within the firing pin impression hole, but also the plateaued
region surrounding the hole that does not come into contact with the breech
face.}
}
\value{
An x3p object containing the surface matrix of a breech face
impression scan where the observations on the interior of the firing pin
impression hole have been filtered out.
}
\description{
Filter-out the firing pin impression region of a breech face impression scan
}
\note{
The radius estimation procedure effectively estimates the radius
of the firing pin hole. Unfortunately, it is often desired that more than
just observations in firing pin hole are removed. In particular, the
plateaued region surrounding the firing pin impression hole does not come
into contact with the breech face of a firearm and is thus unwanted in the
final, processed scan. The radiusOffset argument must be tuned (around 200
seems to work well for the Fadul cartridge cases) to remove these unwanted
observations.
}
\keyword{internal}
