% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_gaussFilter}
\alias{preProcess_gaussFilter}
\title{Performs a low, high, or bandpass Gaussian filter on a surface matrix with a
particular cut-off wavelength.}
\usage{
preProcess_gaussFilter(
  surfaceMat,
  res,
  wavelength = c(16, 500),
  filtertype = "bp"
)
}
\arguments{
\item{surfaceMat}{a surface matrix representing a breech face impression scan}

\item{res}{sampling resolution of the surface matrix}

\item{wavelength}{cut-off wavelength}

\item{filtertype}{specifies whether a low pass, "lp", high pass, "hp", or bandpass,
"bp" filter is to be used. Note that setting filterype = "bp" means that
wavelength should be a vector of two numbers. In this case, the max of
these two number will be used for the high pass filter and the min for the
low pass filter.}
}
\description{
Performs a low, high, or bandpass Gaussian filter on a surface matrix with a
particular cut-off wavelength.
}
\examples{
\dontrun{
raw_x3p <- x3ptools::read_x3p("path/to/file.x3p") \%>\%
x3ptools::sample_x3p(m = 2)

raw_x3p$surface.matrix <- raw_x3p$surface.matrix \%>\%
  cmcR::preProcess_ransac() \%>\%
  cmcR::preProcess_levelBF() \%>\%
  cmcR::preProcess_cropWS() \%>\%
  cmcR::preProcess_removeFPCircle() \%>\%
  cmcR::preProcess_gaussFilter(res = raw_x3p$header.info$incrementY,
                               wavelength = c(16,500),
                               filtertype = "bp")
}

}
\seealso{
https://www.mathworks.com/matlabcentral/fileexchange/61003-filt2-2d-geospatial-data-filter?focused=7181587&tab=example
}
