% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_levelBF}
\alias{preProcess_levelBF}
\title{Given the output of preProcess_ransac, extracts values (either raw or residual) from the surface matrix to which the RANSAC plane was fit.}
\usage{
preProcess_levelBF(ransacFit, useResiduals = FALSE)
}
\arguments{
\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}
}
\description{
Given the output of preProcess_ransac, extracts values (either raw or residual) from the surface matrix to which the RANSAC plane was fit.
}
