% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_removeBFTrend}
\alias{preProcess_removeBFTrend}
\title{Level a breech face impression surface matrix by a conditional statistics}
\usage{
preProcess_removeBFTrend(x3p, statistic = "mean", ...)
}
\arguments{
\item{x3p}{an x3p object containing the surface matrix of a cartridge case
scan}

\item{statistic}{either "mean" or "quantile"}

\item{...}{arguments to be set in the quantreg::rq function if statistic =
"quantile" is set. In this case, tau = .5 and method = "fn" are recommended}
}
\value{
an x3p object containing the leveled cartridge case scan surface
matrix.
}
\description{
Level a breech face impression surface matrix by a conditional statistics
}
\examples{
fadul1.1 <- x3ptools::read_x3p("https://tsapps.nist.gov/NRBTD/Studies/CartridgeMeasurement/DownloadMeasurement/2d9cc51f-6f66-40a0-973a-a9292dbee36d")

fadul1.1_cropped <- preProcess_cropBFExterior(x3p = fadul1.1,radiusOffset = -20)

fadul1.1_filtered <- preProcess_filterBFInterior(x3p = fadul1.1_cropped, radiusOffset = 200)

fadul1.1_leveled <- preProcess_removeBFTrend(x3p = fadul1.1_filtered,statistic = "quantile",tau = .5,method = "fn")

x3pListPlot(list("Original" = fadul1.1,"Cropped" = fadul1.1_cropped,"Cropped & Filtered" = fadul1.1_filtered,"Cropped, Filtered, and Leveled" = fadul1.1_leveled))
}
