% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{removeFiringPin}
\alias{removeFiringPin}
\alias{preProcess_removeFPCircle}
\alias{preProcess_filterInterior}
\title{Remove firing pin}
\usage{
preProcess_removeFPCircle(
  x3p,
  aggregationFunction = mean,
  smootherSize = 2 * round((0.1 * nrow(surfaceMat)/2)) + 1,
  gridSize = 40,
  gridGranularity = 1,
  houghScoreQuant = 0.9
)

preProcess_filterInterior(
  x3p,
  scheme = 3,
  high_connectivity = FALSE,
  tolerance = 0,
  radiusOffset = 0
)
}
\arguments{
\item{x3p}{an x3p object containing the surface matrix of a cartridge case
scan}

\item{aggregationFunction}{function to select initial radius estimate from
those calculated using fpRadiusGridSearch}

\item{smootherSize}{size of average smoother (to be passed to zoo::roll_mean)}

\item{gridSize}{size of grid, centered on the initial radius estimate, to be
used to determine the best fitting circle to the surface matrix via the
Hough transform method}

\item{gridGranularity}{granularity of radius grid used to determine the best
fitting circle to the surface matrix via the Hough transform method}

\item{houghScoreQuant}{quantile cut-off to be used when determining a final
radius estimate using the score values returned by the imager::hough_circle}

\item{scheme}{argument for imager::imgradient}

\item{high_connectivity}{argument for imager::label}

\item{tolerance}{argument for imager::label}

\item{radiusOffset}{number of pixels to add to estimated firing pin hole
radius. This is commonly a positive value (e.g., 200) to not only remove
observations within the firing pin impression hole, but also the plateaued
region surrounding the hole that does not come into contact with the breech
face.}
}
\value{
An x3p object containing the surface matrix of a breech face
impression scan where the observations on the interior of the firing pin
impression hole have been filtered out.
}
\description{
Functions to remove the firing pin circle in the center of a
breech face scan. preProcess_removeFPCircle is a slow yet fully automatic
function for removing the firing pin impression circle that uses the
circular Hough transform after estimating the firing pin circle radius.
preProcess_filterInterior is faster yet semi-automatic in that it estimates
the center and radius of the \emph{hole} in the scan left by the firing pin. The
user must input a positive radiusOffset to remove the rest of the plateaued
firing pin impression region.
}
\note{
imager treats a matrix as its transpose (i.e., x and y axes are
swapped). As such, relative to the original surface matrix, the x and y
columns in the data frame fpImpressionCircle actually correspond to the row
and column indices at which the center of the firing pin impression circle
is estiamted to be.

The radius estimation procedure effectively estimates the radius
of the firing pin hole. Unfortunately, it is often desired that more than
just observations in firing pin hole are removed. In particular, the
plateaued region surrounding the firing pin impression hole does not come
into contact with the breech face of a firearm and is thus unwanted in the
final, processed scan. The radiusOffset argument must be tuned (around 200
seems to work well for the Fadul cartridge cases) to remove these unwanted
observations.
}
\examples{
\dontrun{
raw_x3p <- x3ptools::read_x3p("path/to/file.x3p") \%>\%
x3ptools::sample_x3p(m = 2)

raw_x3p$surface.matrix <- raw_x3p$surface.matrix \%>\%
  cmcR::preProcess_ransacLevel() \%>\%
  cmcR::preProcess_levelBF() \%>\%
  cmcR::preProcess_cropExterior() \%>\%
  cmcR::preProcess_removeFPCircle(aggregationFunction = mean,
                                  smootherSize = 2*round((.1*nrow(surfaceMat)/2)) + 1,
                                  gridGranularity = 1,
                                  houghScoreQuant = .9)
}
fadul1.1 <- x3ptools::read_x3p("https://tsapps.nist.gov/NRBTD/Studies/CartridgeMeasurement/DownloadMeasurement/2d9cc51f-6f66-40a0-973a-a9292dbee36d")

fadul1.1_cropped <- preProcess_cropExterior(x3p = fadul1.1,radiusOffset = -20)

fadul1.1_filtered <- preProcess_filterInterior(x3p = fadul1.1_cropped, radiusOffset = 200)

x3pListPlot(list("Original" = fadul1.1,"Cropped" = fadul1.1_cropped,"Cropped & Filtered" = fadul1.1_filtered))
}
