% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfComputation.R
\name{round2}
\alias{round2}
\title{Copies behavior of round() function in MATLAB}
\usage{
round2(x, n)
}
\description{
Copies behavior of round() function in MATLAB
}
\seealso{
http://stackoverflow.com/questions/12688717/round-up-from-5-in-r
}
\keyword{internal}
