% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{selectBFImpression}
\alias{selectBFImpression}
\title{Select breech face impression from a cartridge case scan}
\usage{
selectBFImpression(
  x3p_path,
  ransacInlierThresh = (10^(-5)),
  ransacFinalSelectThresh = 2 * (10^(-5)),
  ransacIters = 150,
  useResiduals = TRUE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  gaussFilterRes,
  gaussFilterWavelength = c(16, 250),
  gaussFilterType = "bp"
)
}
\arguments{
\item{x3p_path}{path to an .x3p file}

\item{ransacInlierThresh}{threshold to declare an observed value close to the
fitted plane an "inlier" for the RANSAC method}

\item{ransacFinalSelectThresh}{once the RANSAC plane is fitted based on the
inlierThreshold, this argument dictates which observations are selected as
the final breech face estimate.}

\item{ransacIters}{number of candidate planes to fit for the RANSAC method
(higher value yields more stable breech face estimate)}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}

\item{croppingThresh}{treshold for cropping exterior whitespace from the
breech face impression}

\item{standardizeBF}{subtract mean and divide by standard deviation of
processed surface matrix}

\item{gaussFilterRes}{\strong{Optional} sampling resolution of the surface matrix
(given by incrementX or incrementY of the header.info element of an x3p
object)}

\item{gaussFilterWavelength}{\strong{Optional} cut-off wavelength to be attenuated}

\item{gaussFilterType}{\strong{Optional} specifies whether a low pass, "lp", high
pass, "hp", or bandpass, "bp" filter is to be used. Note that setting
filterype = "bp" means that wavelength should be a vector of two numbers.
In this case, the max of these two number will be used for the high pass
filter and the min for the low pass filter.}
}
\value{
x3p object containing the processed breech face
}
\description{
Given a string representing a path to a cartridge case scan .x3p
file, this function will read and process the scan including (1) use the
RANSAC iterative plane fitting method to detect the height value at which
the breech face impressions are in the scan, (2) crop out extra whitespace,
NA values from the exterior of the resulting breech face impression surface
matrix, and (3) detect and filter the inner firing pin impression circle
using a circular Hough transform. Optionally, a Gaussian filter can be
applied. See the cmcR::preProcess_gaussFilter function for more
information.

Note that the RANSAC method is applied twice to the scan: once to get a
rough estimate of the breech face impression height values (controlled with
the ransacInlierThresh and ransacIters arguments) and again using
.1\emph{ransacInlierThresh and 2}ransacIters to get a more precise estimate.
}
\examples{
\dontrun{
 processedBF1 <- cmcR::selectBFImpression(x3p_path = "path/to/file.x3p")
}

}
\seealso{
cartridges3D package (LINK)

x3ptools (LINK)
}
