% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{selectBFImpression_resize}
\alias{selectBFImpression_resize}
\title{Selected breech face impression from a cartridge case scan after using
imager::resize to resize the scan}
\usage{
selectBFImpression_resize(
  x3p_path,
  ransacInlierThresh = (10^(-5)),
  ransacFinalSelectThresh = 2 * (10^(-5)),
  ransacIters = 150,
  useResiduals = FALSE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  size_x,
  size_y = size_x,
  interpolation_type = 1,
  boundary_conditions = 0
)
}
\arguments{
\item{x3p_path}{path to a .x3p file}

\item{ransacInlierThresh}{threshold to declare an observed value close to the
fitted plane an "inlier" for the RANSAC method}

\item{ransacIters}{number of candidate planes to fit for the RANSAC method
(higher value yields more stable breech face estimate)}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}

\item{croppingThresh}{treshold for cropping exterior whitespace from the
breech face impression}

\item{size_x}{Number of columns (new size along the X-axis). Note: the
"X"-axis according to imager is actually the rows of the matrix}

\item{size_y}{Number of rows (new size along the Y-axis). Note: the "Y"-axis
according to imager is actually the number of cols of the matrix}

\item{interpolation_type}{Method of interpolation: -1 = no interpolation: raw
memory resizing. 0 = no interpolation: additional space is filled according
to boundary_conditions. 1 = nearest-neighbor interpolation. 2 = moving
average interpolation. 3 = linear interpolation. 4 = grid interpolation. 5
= cubic interpolation. 6 = lanczos interpolation.}

\item{boundary_conditions}{Border condition type.}

\item{standardizedBF}{subtract mean and divide by standard deviation of
processed surface matrix}
}
\value{

}
\description{
Selected breech face impression from a cartridge case scan after using
imager::resize to resize the scan
}
\note{
imager treats a matrix as its transpose (i.e., x and y axes are
swapped). As such the size_x argument corresponds to changing the number of
rows in the original matrix and size_y the number of columns.
}
\examples{
\dontrun{

}
}
\seealso{
cartridges3D package (LINK)

imager package (LINK)
}
