% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{selectBFImpression_sample_x3p}
\alias{selectBFImpression_sample_x3p}
\title{Implements x3ptools::sample_x3p to downsample an x3p object's surface matrix.}
\usage{
selectBFImpression_sample_x3p(
  x3p_path,
  ransacInlierThresh = (10^(-5)),
  ransacFinalSelectThresh = 2 * (10^(-5)),
  ransacIters = 150,
  useResiduals = FALSE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  m = 2,
  mY = m,
  offset = 0,
  offsetY = offset,
  ...
)
}
\arguments{
\item{x3p_path}{path to a .x3p file}

\item{ransacInlierThresh}{threshold to declare an observed value close to the
fitted plane an "inlier" for the RANSAC method}

\item{ransacIters}{number of candidate planes to fit for the RANSAC method
(higher value yields more stable breech face estimate)}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}

\item{croppingThresh}{treshold for cropping exterior whitespace from the
breech face impression}

\item{m}{integer value - every mth value is included in the sample}

\item{mY}{integer value - every mth value is included in the sample in the x
direction and every mYth value is included in y direction}

\item{offset}{integer value between 0 and m-1 to specify offset of the sample}

\item{offsetY}{integer value between 0 and mY-1 to specify different offsets
for x and y direction}

\item{...}{arguments to be passed to fpCircleHoughDetection function. This is
useful if, for example, the Hough transform appears to be mis-estimating
where the firing pin impression circle is in the scan.}

\item{finalSelectionThreshold}{once the RANSAC plane is fitted based on the
inlierThreshold, this argument dictates which observations are selected as
the final breech face estimate.}

\item{standardizedBF}{subtract mean and divide by standard deviation of
processed surface matrix}
}
\value{

}
\description{
Implements x3ptools::sample_x3p to downsample an x3p object's surface matrix.
}
\note{
Given a matrix, x3ptools populates an x3p object's surface matrix
starting in the top left corner moving right by reading the the matrix from
the bottom left corner and moving up. Effectively rotating the matrix by 90
degrees clockwise.
}
\examples{
\dontrun{
    set.seed(0)
    testImage <- findPlaneRansac(surfaceMat)
}
}
\seealso{
cartridges3D package (LINK)
}
