% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{shiftSurfaceMatbyMean}
\alias{shiftSurfaceMatbyMean}
\title{Subtract the average pixel value from an image and replace NAs with 0}
\usage{
shiftSurfaceMatbyMean(surfaceMat, m)
}
\arguments{
\item{surfaceMat}{surface matrix}

\item{mean}{average pixel value to shift an image by}
}
\description{
This is a helper for the cellCCF function.
}
