% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_old.R
\name{standardizeSurfaceMat}
\alias{standardizeSurfaceMat}
\title{Subtract the average pixel value from an image and replace NAs with 0}
\usage{
standardizeSurfaceMat(surfaceMat, m, s)
}
\arguments{
\item{surfaceMat}{surface matrix}

\item{mean}{average pixel value to shift an image by}
}
\description{
This is a helper for the cellCCF function.
}
\keyword{internal}
