% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{topResultsPerCell}
\alias{topResultsPerCell}
\title{Calculate dx, dy, theta values at which top similarity is achieved in a cell
pair.}
\usage{
topResultsPerCell(ccfResults)
}
\arguments{
\item{ccfResults}{list of data frames, like the one returned by
cmcR::cellCCF, containing breech face cell comparison results}
}
\description{
Given a list of data frames like the one returned in the
\code{ccfResults} element of the list returned from cmcR::cellCCF, returns a
data frame of the theta, dx, dy, ccf, and rawCorr values at which each cell
pair attains maximum rawCorr.
}
\examples{
\dontrun{
#x3p1 and x3p2 are two x3p objects containing processed surface matrices
comparison1 <- cellCCF(x3p1,x3p2) #defaults assumed

comparison1$ccfResults \%>\%
  cmcR::topResultsPerCell()
}

}
\seealso{
cmcR::cellCCF
}
