% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{ccfMapPlot}
\alias{ccfMapPlot}
\title{Plots the CCF map between two matrices}
\usage{
ccfMapPlot(mat1, mat2, returnGrob = FALSE)
}
\arguments{
\item{mat1}{a matrix}

\item{mat2}{another matrix}

\item{returnGrob}{if TRUE, then function will return the gridExtra grob
object}
}
\description{
Uses the gridExtra::arrangeGrob function to put 3 plots and a
table on the same grob object.
}
