% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{cellCCF_bothDirections}
\alias{cellCCF_bothDirections}
\title{Wrapper for applying the cmcR::cellCCF function to x3p1 vs. x3p2 and again
for x3p2 vs. x3p1. See cellCCF function documentation for more details.}
\usage{
cellCCF_bothDirections(
  x3p1,
  x3p2,
  thetas = seq(-30, 30, by = 3),
  cellNumHoriz = 7,
  cellNumVert = cellNumHoriz,
  centerCell,
  scaleCell,
  ...
)
}
\description{
Wrapper for applying the cmcR::cellCCF function to x3p1 vs. x3p2 and again
for x3p2 vs. x3p1. See cellCCF function documentation for more details.
}
\seealso{
cellCCF function
}
