% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{cellCCF_bothDirections}
\alias{cellCCF_bothDirections}
\title{Applies cellCCF function twice for a pair of cartridge case scans.}
\usage{
cellCCF_bothDirections(
  x3p1,
  x3p2,
  thetas = seq(-30, 30, by = 3),
  cellNumHoriz = 7,
  regionToCellProp = 4,
  cellNumVert = cellNumHoriz,
  centerCell,
  scaleCell,
  ...
)
}
\arguments{
\item{x3p1}{(no default) an x3p object containing the surface matrix of a
breech face impression}

\item{x3p2}{(no default) an x3p object containing the surface matrix of a
breech face impression to be compared to that in x3p1}

\item{thetas}{(default seq(from = -30,to = 30,by = 3)) rotation values (in
degrees) for which x3p2$surface.matrix will be rotated, split into cells,
and compared to x3p1$surface.matrix}

\item{cellNumHoriz}{(default 7) number of cells along horizontal axis to
divide x3p1$surface.matrix into}

\item{regionToCellProp}{determines how much larger the x3p2 regions will be
relative to the x3p1 cells. For example, if regionToCellProp = 4 means that
the x3p2 regions will be 4 times times larger (sidelengths multiplied by 2)}

\item{cellNumVert}{(default equal to cellNumHoriz) number of cells along
vertical axis to divide x3p1$surface.matrix into}

\item{centerCell}{\strong{OPTIONAL} (default missing) dictates if cell is to be
shifted prior to CCF calculation. Default is that no shifting is performed.
If set to "wholeMatrix", then each cell is subracted by the mean of the
whole matrix. If set to "individualCell", then each cell is subtracted by
its particular mean.}

\item{scaleCell}{\strong{OPTIONAL} (default missing) dictates if cell is to be
scaled prior to CCF calculation. Default is that no scaling is performed.
If set to "wholeMatrix", then each cell is divided by the standard
deviation of the whole matrix. If set to "individualCell", then each cell
is divided by its particular standard deviation.}

\item{minObservedProp}{(default 15) the minimum proportion of a cell that
needs to contain observed (i.e., non-NA) values for it to be included in
the CCF calculation procedure}
}
\description{
Wrapper for applying the cmcR::cellCCF function to x3p1 vs. x3p2 and again
for x3p2 vs. x3p1. See cellCCF function documentation for more details.
}
\examples{
\dontrun{
 #x3p1 and x3p2 are assumed to be 2 processed scans
comparison1 <- cellCCF_bothDirections(x3p1 = x3p1,
                                      x3p2 = x3p2,
                                      thetas = seq(-30,30,by = 30),
                                      cellNumHoriz = 8,
                                      regionToCellProp = 4,
                                      centerCell = "individualCell",
                                      scaleCell = "individualCell")

comparison1$comparison_1to2 #comparison of x3p1 vs x3p2
comparison1$comparison_2to1 #comparison of x3p2 vs x3p1
}

}
\seealso{
cmcR::cellCCF_improved
}
