% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{cellDivision}
\alias{cellDivision}
\title{Divides an image into cells}
\usage{
cellDivision(surfaceMat, cellNumHoriz = 8, cellNumVert = 8)
}
\arguments{
\item{surfaceMat}{surface matrix of a breech face impression}

\item{cellNumHoriz}{number of splits along horizontal axis}

\item{cellNumVert}{number of splits along vertical axis}
}
\description{
This is a helper for the cellCCF function.
}
\keyword{internal}
