% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{cmcFilter}
\alias{cmcFilter}
\title{Implements CMC logic on a single data frame as proposed by Song (2013).}
\arguments{
\item{ccfDF}{data frame containing ccf results from a comparison between two
cartridge case scans}

\item{consensus_function}{function to aggregate the translation (dx and dy)
and rotation (theta) values in the ccfDF data frame to determine
"consensus" values}

\item{ccf_thresh}{minimum correlation threshold to call a cell pair
"congruent matching"}

\item{dx_thresh}{maximum distance from the consensus dx value that a cell
pair can be to be called "congruent matching"}

\item{dy_thresh}{maximum distance from the consensus dy value that a cell
pair can be to be called "congruent matching"}

\item{theta_thresh}{maximum distance from the consensus theta value that a
cell pair can be to be called "congruent matching"}

\item{consensus_function_theta}{\emph{(OPTIONAL)} function (separate from
consensus_function) to aggregate the rotation (theta) values in the ccfDF
data frame to determine "consensus" values}

\item{...}{arguments to be passed to consensus_function or
consensus_function_theta (e.g., na.rm = TRUE) if necessary}
}
\description{
Applies initially proposed Congruent Matching Cells method logic
to the CCF results of a comparison between two cartridge case scans.
}
\examples{
\dontrun{
#x3p1 and x3p2 are two x3p objects containing processed surface matrices
comparison1 <- cellCCF(x3p1,x3p2) #defaults assumed

#calculate "initial" cmcs
comparison1$ccfResults \%>\%
  cmcR::topResultsPerCell() \%>\%
  cmcR::cmcFilter(consensus_function = median,
                  ccf_thresh = .4,
                  dx_thresh = 20,
                  dy_thresh = dx_thresh,
                  theta_thresh = 3,
                  consensus_function_theta = median)
}

}
\seealso{
\url{https://pdfs.semanticscholar.org/4bf3/0b3a23c38d8396fa5e0d116cba63a3681494.pdf}
}
