% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{cmcFilter_improved}
\alias{cmcFilter_improved}
\title{Implements "improved" CMC logic on a list of CCF results for a comparison
between two cartridge case scans as proposed by Tong et al. (2015)}
\usage{
cmcFilter_improved(
  cellCCF_bothDirections_output,
  consensus_function = median,
  corr_thresh = 0.4,
  dx_thresh = 20,
  dy_thresh = dx_thresh,
  theta_thresh = 3,
  consensus_function_theta = consensus_function,
  ...
)
}
\arguments{
\item{cellCCF_bothDirections_output}{list returned by the function
cmcR::cellCCF_bothdirections}

\item{consensus_function}{function to aggregate the translation (dx and dy)
and rotation (theta) values in the ccfDF data frame to determine
"consensus" values}

\item{corr_thresh}{minimum correlation threshold to call a cell pair
"congruent matching"}

\item{dx_thresh}{maximum distance from the consensus dx value that a cell
pair can be to be called "congruent matching"}

\item{dy_thresh}{maximum distance from the consensus dy value that a cell
pair can be to be called "congruent matching"}

\item{theta_thresh}{maximum distance from the consensus theta value that a
cell pair can be to be called "congruent matching"}

\item{consensus_function_theta}{\emph{(OPTIONAL)} function (separate from
consensus_function) to aggregate the rotation (theta) values in the ccfDF
data frame to determine "consensus" values}

\item{...}{arguments to be passed to consensus_function or
consensus_function_theta (e.g., na.rm = TRUE) if necessary}
}
\description{
Implements "improved' Congruent Matching Cells logic, as
proposed by Tong et al. (2015), to the CCF results of a comparison between
two cartridge case scans.
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4730689/pdf/jres.120.008.pdf}
}
