% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{cmcPlot}
\alias{cmcPlot}
\title{Visualize initial and final CMCs for a cartridge case pair comparison}
\usage{
cmcPlot(x3p1, x3p2, cellCCF_bothDirections_output, cmcFilter_improved_output)
}
\arguments{
\item{x3p1}{an x3p object}

\item{x3p2}{a different x3p object}

\item{cellCCF_bothDirections_output}{output from the function cmcR::cellCCF_bothDirections}

\item{cmcFilter_improved_output}{output from the function cmcR::cmcFilter_improved}
}
\description{
Visualize initial and final CMCs for a cartridge case pair comparison
}
