% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{cmcPlot}
\alias{cmcPlot}
\title{Plots the Congruent Matching Cells identified in a cartridge case scan}
\usage{
cmcPlot(x3p, cmcDF, method = "cimg")
}
\arguments{
\item{x3p}{The "questioned" cartridge case scan, in the form of an x3p
object, for which CMCs were determined (this would be x3p1 in a call to
cellCCF)}

\item{cmcDF}{data frame containing the congruent matching cells of the
questioned x3p object}

\item{method}{dictates whether a cimg plot is returned, or a ggplot2 plot.
The ggplot2 plot looks nicer, but takes much longer to construct than the
cimg plot.}
}
\description{
Plots the selected congruent matching cells of a questioned x3p by either
converting the surface matrix contained in the x3p to a cimg object and
using base plot (fast, but doesn't look particularly great) or by plotting
surface matrix using ggplot2::geom_tile (slow, but nice looking).
}
\examples{
\dontrun{
comparison1 <- cellCCF_bothDirections(x3p1,x3p2)

cmcs <- cmcFilter_improved(comparison1)

#initially selected CMCs based on x3p1 vs x3p2 comparison
cmcPlot(x3p1,cmcs$initialCMCs[[1]][[1]])

#initialy selected CMCs based on x3p2 vs x3p1
cmcPlot(x3p2,cmcs$initialCMCs[[1]][[2]])

#final selected CMCs (may be empty) based on both x3p1 vs x3p2 and x3p2 vs x3p1
cmcPlot(x3p1,cmcs$finalCMCs)

#make plot using ggplot2::geom_tile
cmcPlot(x3p1,cmcs$finalCMCs,method = "ggplot2")
}

}
