% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{cmcPlot}
\alias{cmcPlot}
\title{Visualize initial and high CMCs for a cartridge case pair comparison}
\arguments{
\item{referenceScan}{an x3p object}

\item{targetScan}{a different x3p object}

\item{reference_v_target_CMCs}{CMCs for the comparison between the reference
scan and the target scan.}

\item{target_v_reference_CMCs}{(optional) CMCs for the comparison between the
target scan and the reference scan. If this is missing, then only the
original method CMCs will be plotted}

\item{type}{argument to be passed to cmcR::x3pListPlot function}

\item{x3pNames}{(Optional) Names of x3p objects to be included in x3pListPlot
function}

\item{legend.quantiles}{vector of quantiles to be shown as tick marks on
legend plot}

\item{height.colors}{vector of colors to be passed to scale_fill_gradientn
that dictates the height value colorscale}

\item{cell.colors}{vector of 2 colors for plotting non-matching and matching
(in that order) cells}

\item{na.value}{color to be used for NA values (passed to
scale_fill_gradientn)}
}
\description{
Constructs either a single faceted plot or a list of plots
depicting the CMCs/non-CMCs under the initially proposed and High CMC
methods for a pair of cartridge case scans
}
\examples{
\dontrun{
comparison <- cellCCF_bothDirections(referenceScan,targetScan)
cmcs <- cmcFilter_improved(comparison)

cmcPlot(referenceScan,targetScan,comparison,cmcs,type = "faceted",x3pNames = c("name1","name2"))
}

}
