% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_allTogether}
\alias{comparison_allTogether}
\title{Performs all steps in the cell-based comparison procedure.}
\usage{
comparison_allTogether(
  reference,
  target,
  theta = 0,
  numCells = 64,
  maxMissingProp = 0.85
)
}
\description{
Performs all steps in the cell-based comparison procedure.
}
