% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_calcPropMissing}
\alias{comparison_calcPropMissing}
\title{Calculate the proportion of missing values in a surface matrix}
\usage{
comparison_calcPropMissing(heightValues)
}
\description{
Calculate the proportion of missing values in a surface matrix
}
