% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_cellDivision}
\alias{comparison_cellDivision}
\title{Split a reference scan into a grid of cells}
\usage{
comparison_cellDivision(x3p, numCells = 64)
}
\description{
Split a reference scan into a grid of cells
}
