% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_cor}
\alias{comparison_cor}
\title{Calculates correlation between a cell and a matrix of the same dimensions
extracted from the cell's associated region.}
\usage{
comparison_cor(
  cellHeightValues,
  regionHeightValues,
  fft_ccf_df,
  use = "pairwise.complete.obs"
)
}
\description{
Calculates correlation between a cell and a matrix of the same dimensions
extracted from the cell's associated region.
}
\seealso{
\url{(https://mathworld.wolfram.com/Cross-CorrelationTheorem.html)}
}
