% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_fft_ccf}
\alias{comparison_fft_ccf}
\title{Estimate translation alignment between a cell/region pair based on the
Cross-Correlation Theorem.}
\usage{
comparison_fft_ccf(cellHeightValues, regionHeightValues)
}
\description{
Estimate translation alignment between a cell/region pair based on the
Cross-Correlation Theorem.
}
\note{
The FFT is not defined for matrices containing missing values. The
missing values in the cell and region need to be replaced before using this
function. See the \link[cmcR]{comparison_standardizeHeights} function
to replace missing values after standardization.
}
