% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_ccf.R
\name{comparison_getTargetRegions}
\alias{comparison_getTargetRegions}
\title{Extract regions from a target scan based on associated cells in reference
scan}
\usage{
comparison_getTargetRegions(
  cellHeightValues,
  cellRange,
  target_x3p,
  rotation = 0,
  regionSizeMultiplier = 9
)
}
\description{
Extract regions from a target scan based on associated cells in reference
scan
}
