% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{cropScanWhitespace}
\alias{cropScanWhitespace}
\title{Crop out rows/columns outside of the breech face impression in a cartridge
case scan}
\usage{
cropScanWhitespace(surfaceMat, croppingThreshold = 10, ...)
}
\description{
Crop out rows/columns outside of the breech face impression in a cartridge
case scan
}
