% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{cropScanWhitespace}
\alias{cropScanWhitespace}
\title{Crop out rows/columns outside of the breech face impression in a cartridge
case scan. Helper for selectBFImpression}
\usage{
cropScanWhitespace(surfaceMat, croppingThresh = 1, ...)
}
\description{
Crop out rows/columns outside of the breech face impression in a cartridge
case scan. Helper for selectBFImpression
}
