% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{decision_highCMC_classifyCMCs}
\alias{decision_highCMC_classifyCMCs}
\title{Apply CMC classification logic of the Tong et al. (2015) to the CMC-theta
distribution returned by the decision_highCMC_cmcThetaDistrib function}
\usage{
decision_highCMC_classifyCMCs(
  comparisonFeaturesDF,
  corColName = "pairwiseCompCor",
  xThresh = 20,
  yThresh = xThresh,
  thetaThresh = 6,
  corThresh = 0.5,
  tau = 1
)
}
\description{
Apply CMC classification logic of the Tong et al. (2015) to the CMC-theta
distribution returned by the decision_highCMC_cmcThetaDistrib function
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4730689/pdf/jres.120.008.pdf}
}
