% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{decision_highCMC_classifyCMCs}
\alias{decision_highCMC_classifyCMCs}
\title{Apply CMC classification logic of the Tong et al. (2015) to the CMC-theta
distribution returned by the decision_highCMC_cmcThetaDistrib function}
\usage{
decision_highCMC_classifyCMCs(
  cellIndex,
  x,
  y,
  theta,
  corr,
  xThresh = 20,
  yThresh = xThresh,
  thetaThresh = 6,
  corrThresh = 0.5,
  tau = 1
)
}
\arguments{
\item{cellIndex}{vector/tibble column containing cell indices corresponding to
a reference cell}

\item{x}{vector/tibble column containing x horizontal translation values}

\item{y}{vector/tibble column containing y vertical translation values}

\item{theta}{vector/tibble column containing theta rotation values}

\item{corr}{vector/tibble column containing correlation similarity scores
between a reference cell and its associated target region}

\item{xThresh}{used to classify particular x values "congruent" (conditional
on a particular theta value) if they are within xThresh of the
theta-specific median x value}

\item{yThresh}{used to classify particular y values "congruent" (conditional
on a particular theta value) if they are within yThresh of the
theta-specific median y value}

\item{thetaThresh}{defines how wide a High CMC mode is allowed to be in the
CMC-theta distribution before it's considered too diffuse}

\item{corrThresh}{to classify particular correlation values "congruent"
(conditional on a particular theta value) if they are at least corrThresh}

\item{tau}{constant used to define a "high" CMC count. This number is
subtracted from the maximum CMC count achieved in the CMC-theta
distribution. Theta values with CMC counts above this value are considered
to have "high" CMC counts.}
}
\description{
Apply CMC classification logic of the Tong et al. (2015) to the CMC-theta
distribution returned by the decision_highCMC_cmcThetaDistrib function
}
\note{
the decision_CMC function internally calls this function if a value of
tau is provided
}
\seealso{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4730689/pdf/jres.120.008.pdf}
}
\keyword{internal}
