% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcLogic.R
\name{decision_highCMC_cmcThetaDistrib}
\alias{decision_highCMC_cmcThetaDistrib}
\title{Compute CMC-theta distribution for a set of comparison features}
\usage{
decision_highCMC_cmcThetaDistrib(
  cellIndex,
  x,
  y,
  theta,
  corr,
  corColName = "pairwiseCompCor",
  xThresh = 20,
  yThresh = xThresh,
  corrThresh = 0.5
)
}
\description{
Compute CMC-theta distribution for a set of comparison features
}
