% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{findPlaneRansac}
\alias{findPlaneRansac}
\title{Finds plane of breechface marks using RANSAC}
\usage{
findPlaneRansac(surfaceMat, inlierTreshold = 10^(-5), iters = 75, ...)
}
\arguments{
\item{surfaceMat}{matrix of input depths in microns.}
}
\value{
List object containing fitted plane and selected breechface marks.
}
\description{
Given input depths (in microns), find best-fitting plane using RANSAC. This
should be the plane that the breechface marks are on. Adapted from cartridges3D::findPlaneRansac function
}
\examples{
\dontrun{
    set.seed(0)
    testImage <- findPlaneRansac(surfaceMat)
}
}
\seealso{
cartridges3D package (LINK)
}
