% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{levelBFImpression}
\alias{levelBFImpression}
\title{Helper for selectBFImpression. This ia modified version of the levelBF3D
function available in the cartridges3D package on GitHub.}
\usage{
levelBFImpression(ransacFit, useResiduals = TRUE, ...)
}
\arguments{
\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}
}
\description{
Helper for selectBFImpression. This ia modified version of the levelBF3D
function available in the cartridges3D package on GitHub.
}
\seealso{
https://github.com/xhtai/cartridges3D
}
\keyword{internal}
