% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{preProcess_crop}
\alias{preProcess_crop}
\title{Removes observations from a surface matrix on the exterior of interior of the
cartridge case surface.}
\usage{
preProcess_crop(
  x3p,
  region = "exterior",
  radiusOffset = 0,
  croppingThresh = 1,
  agg_function = median,
  scheme = 3,
  high_connectivity = FALSE,
  tolerance = 0
)
}
\arguments{
\item{region}{dictates whether the observations on the "exterior" or
"interior" of the scan are removed}

\item{radiusOffset}{the procedures used to detect the interior/exterior of
the scan tend to under/overestimate the radius of the region of interest,
respectively. The number passed to this argument is added to the radius
estimate as an offset.}
}
\description{
Removes observations from a surface matrix on the exterior of interior of the
cartridge case surface.
}
