% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess_detectFPCircle}
\alias{preProcess_detectFPCircle}
\title{Detect the radius and center of a firing pin impression circle in a breech
face impression scan using a circular Hough transform.}
\usage{
preProcess_detectFPCircle(
  surfaceMat,
  aggregationFunction = mean,
  smootherSize = 2 * round((0.1 * nrow(surfaceMat)/2)) + 1,
  gridSize = 40,
  gridGranularity = 1,
  houghScoreQuant = 0.9
)
}
\arguments{
\item{aggregationFunction}{function to aggregate all 12 radius estimates
determined under the initial radius estimation procedure those calculated
using fpRadiusGridSearch}

\item{smootherSize}{size of average smoother (to be passed to zoo::roll_mean)
used to determine where the non-NA pixel count-per-row series attains a
mode.}

\item{gridGranularity}{granularity of radius grid used to determine the best
fitting circle to the surface matrix via the Hough transform method}

\item{houghScoreQuant}{quantile cut-off to be used when determining a final
radius estimate using the score values returned by the imager::hough_circle
function}

\item{x3p}{an x3p object containing a surface matrix}
}
\description{
This function estimates the radius of a firing pin impression
within a breech face impression scan. It does so by detecting local maxima
in the non-NA pixel count by row/col within the scan. To make the algorithm
more robust, multiple radii estimates are considered by rotating the image
15, 30, 45, 60, and 75 degrees and again detecting local maxima in the
non-NA pixel count by row/col. Based on the argument passed to
aggregationFunction, these radii estimates are reduced to a single, rough
radius estimate (e.g., minimum was determined to be an effective
aggregation function in preliminary tests). A grid of radii values centered
on this estimate are then tested to determine which a final estimate. The
grid mesh size is determined by the argument gridGranularity. A hough transform is
applied to the breech face impression scan for each radius value in the
grid. A final estimate is determined by finding the longest consecutive
sequence of radii values with high associated hough scores. How we
determine "high" hough scores is determined by the houghScoreQuant
argument. Once the longest sequence of high hough score radii values is
found, the average of these radii values is used as the final radius
estimate.
}
\keyword{internal}
