% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_gaussFilter}
\alias{preProcess_gaussFilter}
\title{Performs a low, high, or bandpass Gaussian filter on a surface matrix with a
particular cut-off wavelength.}
\usage{
preProcess_gaussFilter(
  surfaceMat,
  res = selectedBF_x3p$header.info$incrementY,
  wavelength,
  filtertype = "bp"
)
}
\arguments{
\item{surfaceMat}{a surface matrix representing a breech face impression scan}

\item{res}{sampling resolution of the surface matrix}

\item{wavelength}{cut-off wavelength}
}
\description{
Performs a low, high, or bandpass Gaussian filter on a surface matrix with a
particular cut-off wavelength.
}
\seealso{
https://www.mathworks.com/matlabcentral/fileexchange/61003-filt2-2d-geospatial-data-filter?focused=7181587&tab=example
}
