% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess_levelBF}
\alias{preProcess_levelBF}
\title{Levels a breech face impression matrix basedo on a RANSAC-fitted plane}
\usage{
preProcess_levelBF(ransacFit, useResiduals = TRUE)
}
\arguments{
\item{ransacFit}{output from the cmcR::preProcess_ransacLevel function.}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}
}
\value{
a surface matrix of either "raw" breech face values that are inliers
to the RANSAC-fitted plane or residuals between the fitted plane and
observed values.
}
\description{
Given the output of preProcess_ransacLevel, extracts values
(either raw or residual) from the surface matrix to which the RANSAC plane
was fit. Adapted from the cartridges3D::levelBF3D function. This ia
modified version of the levelBF3D function available in the cartridges3D
package on GitHub.
}
\examples{
\dontrun{
raw_x3p <- x3ptools::read_x3p("path/to/file.x3p") \%>\%
  x3ptools::sample_x3p(m = 2)

raw_x3p$surface.matrix <- raw_x3p$surface.matrix \%>\%
 cmcR::preProcess_ransacLevel() \%>\%
 cmcR::preProcess_levelBF(useResiduals = TRUE)
}

}
\seealso{
https://github.com/xhtai/cartridges3D
}
\keyword{internal}
