% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_ransac}
\alias{preProcess_ransac}
\title{Finds plane of breechface marks using RANSAC}
\usage{
preProcess_ransac(
  surfaceMat,
  inlierTreshold = (10^(-5)),
  finalSelectionThreshold = 2 * (10^(-5)),
  iters = 150
)
}
\arguments{
\item{surfaceMat}{a surface matrix representing a breech face impression scan}

\item{finalSelectionThreshold}{once the RANSAC plane is fitted based on the
inlierThreshold, this argument dictates which observations are selected as
the final breech face estimate.}

\item{iters}{number of candidate planes to fit (higher value yields more
stable breech face estimate)}

\item{inlierThreshold}{threshold to declare an observed value close to the
fitted plane an "inlier". A smaller value will yield a more stable
estimate.}
}
\value{
List object containing fitted plane (as an lm object) and selected
breechface marks (matrix of same size as original matrix containing all
inliers close to fitted plane).
}
\description{
Given input depths (in microns), find best-fitting plane using RANSAC. This
should be the plane that the breechface marks are on. Adapted from
cartridges3D::findPlaneRansac function
}
\examples{
\dontrun{
    testImage <- preProcess_ransac(surfaceMat)
}

}
\seealso{
cartridges3D package (LINK)
}
