% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_removeFPCircle}
\alias{preProcess_removeFPCircle}
\title{Given a surface matrix, estimates and filters any pixels within the estimated
firing pin impression circle}
\usage{
preProcess_removeFPCircle(
  surfaceMat,
  aggregation_function = mean,
  smootherSize = 2 * round((0.1 * nrow(surfaceMat)/2)) + 1,
  meshSize = 1,
  houghScoreQuant = 0.9
)
}
\arguments{
\item{surfaceMat}{a surface matrix representing a breech face impression scan}

\item{aggregation_function}{function to select initial radius estimate from
those calculated using fpRadiusGridSearch}

\item{smootherSize}{size of average smoother (to be passed to zoo::roll_mean)}

\item{meshSize}{size of radius mesh to be used for further refinement of the
radius estimate obtained from fpRadiusGridSearch}

\item{houghScoreQuant}{quantile cut-off to be used when determining a final
radius estimate using the score values returned by the imager::hough_circle}
}
\description{
Given a surface matrix, estimates and filters any pixels within the estimated
firing pin impression circle
}
\note{
imager treats a matrix as its transpose (i.e., x and y axes are
swapped). As such, relative to the original surface matrix, the x and y
columns in the data frame fpImpressionCircle actually correspond to the row
and column indices at which the center of the firing pin impression circle
is estiamted to be.
}
