% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessFunctions_modularized.R
\name{preProcess_removeFPCircle}
\alias{preProcess_removeFPCircle}
\title{Given a surface matrix and the output of preProcess_detectFPCircle, filters
any pixels within the estimated firing pin impression circle}
\usage{
preProcess_removeFPCircle(surfaceMat, fpImpressionCircle)
}
\arguments{
\item{surfaceMat}{a surface matrix representing a breech face impression scan}

\item{fpImpressionCircle}{data frame containing 3 columns: "x" and "y"
containing the estimated center of the firing pin impression circle and "r"
containing the estimated radius}
}
\description{
Given a surface matrix and the output of preProcess_detectFPCircle, filters
any pixels within the estimated firing pin impression circle
}
\note{
imager treats a matrix as its transpose (i.e., x and y axes are
swapped). As such, relative to the original surface matrix, the x and y
columns in the data frame fpImpressionCircle actually correspond to the row
and column indices at which the center of the firing pin impression circle
is estiamted to be.
}
