% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{selectBFImpression}
\alias{selectBFImpression}
\title{Selected breech face impression from a cartridge case scan}
\usage{
selectBFImpression(
  x3p_path,
  ransacInlierThresh = (10^(-5)),
  ransacFinalSelectThresh = 2 * (10^(-5)),
  ransacIters = 150,
  useResiduals = FALSE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  preProcess = TRUE
)
}
\arguments{
\item{x3p_path}{path to a .x3p file}

\item{ransacInlierThresh}{threshold to declare an observed value close to the
fitted plane an "inlier" for the RANSAC method}

\item{ransacIters}{number of candidate planes to fit for the RANSAC method
(higher value yields more stable breech face estimate)}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}

\item{croppingThresh}{treshold for cropping exterior whitespace from the
breech face impression}

\item{preProcess}{if FALSE, then no pre-processing is performed. Equivalent
to calling x3ptools::read_x3p(x3p_path)}

\item{finalSelectionThreshold}{once the RANSAC plane is fitted based on the
inlierThreshold, this argument dictates which observations are selected as
the final breech face estimate.}

\item{standardizedBF}{subtract mean and divide by standard deviation of
processed surface matrix}
}
\value{
x3p object containing the processed breech face
}
\description{
Given a string representing a path to a cartridge case scan .x3p file, this function will read and process the scan including (1) use the RANSAC iterative plane fitting method to detect the height value at which the breech face impressions are in the scan, (2) crop out extra whitespace, NA values from the exterior of the resulting breech face impression surface matrix, and (3) detect and filter the inner firing pin impression circle using a circular Hough transform. Note that the RANSAC method is applied twice to the scan: once to get a rough estimate of the breech face impression height values (controlled with the ransacInlierThresh and ransacIters arguments) and again using .1\emph{ransacInlierThresh and 2}ransacIters to get a more precise estimate.
}
\examples{
\dontrun{
 processedBF1 <- cmcR::selectBFImpression(x3p_path = "path/to/file.x3p")
}

}
\seealso{
cartridges3D package (LINK)

x3ptools (LINK)
}
