% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{selectBFImpression_resize}
\alias{selectBFImpression_resize}
\title{Select breech face impression from a cartridge case scan after using
imager::resize to resize the scan.}
\usage{
selectBFImpression_resize(
  x3p_path,
  ransacInlierThresh = (10^(-5)),
  ransacFinalSelectThresh = 2 * (10^(-5)),
  ransacIters = 150,
  useResiduals = FALSE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  size_x,
  size_y = size_x,
  interpolation_type = 1,
  boundary_conditions = 0,
  gaussFilterRes,
  gaussFilterWavelength,
  gaussFilterType
)
}
\arguments{
\item{x3p_path}{path to an .x3p file}

\item{ransacFinalSelectThresh}{threshold to declare an observed value close to the
fitted plane an "inlier" for the RANSAC method}

\item{ransacIters}{number of candidate planes to fit for the RANSAC method
(higher value yields more stable breech face estimate)}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}

\item{croppingThresh}{treshold for cropping exterior whitespace from the
breech face impression}

\item{standardizeBF}{subtract mean and divide by standard deviation of
processed surface matrix}

\item{size_x}{Number of columns (new size along the X-axis). Note: the
"X"-axis according to imager is actually the rows of the matrix}

\item{size_y}{Number of rows (new size along the Y-axis). Note: the "Y"-axis
according to imager is actually the number of cols of the matrix}

\item{interpolation_type}{Method of interpolation: -1 = no interpolation: raw
memory resizing. 0 = no interpolation: additional space is filled according
to boundary_conditions. 1 = nearest-neighbor interpolation. 2 = moving
average interpolation. 3 = linear interpolation. 4 = grid interpolation. 5
= cubic interpolation. 6 = lanczos interpolation.}

\item{boundary_conditions}{Border condition type.}

\item{gaussFilterRes}{\strong{Optional} sampling resolution of the surface matrix
(given by incrementX or incrementY of the header.info element of an x3p
object)}

\item{gaussFilterWavelength}{\strong{Optional} cut-off wavelength to be attenuated}

\item{gaussFilterType}{\strong{Optional} specifies whether a low pass, "lp", high
pass, "hp", or bandpass, "bp" filter is to be used. Note that setting
filterype = "bp" means that wavelength should be a vector of two numbers.
In this case, the max of these two number will be used for the high pass
filter and the min for the low pass filter.}
}
\value{

}
\description{
Select breech face impression from a cartridge case scan after using
imager::resize to resize the scan.
}
\note{
imager treats a matrix as its transpose (i.e., x and y axes are
swapped). As such the size_x argument corresponds to changing the number of
rows in the original matrix and size_y the number of columns.
}
\examples{
\dontrun{
 #use moving average interpolation to resize matrix to be 560x560
 processedBF1 <- cmcR::selectBFImpression_resize(x3p_path = "path/to/file.x3p",size_x = 560,size_y = 560,interpolation_type = 2)
}
}
\seealso{
cartridges3D package (LINK)

imager package (LINK)
}
