% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breechFaceSelection.R
\name{selectBFImpression_sample_x3p}
\alias{selectBFImpression_sample_x3p}
\title{Select breech face impression from a cartridge case scan after using
x3ptools::sample_x3p to downsample the scan.}
\usage{
selectBFImpression_sample_x3p(
  x3p_path,
  ransacInlierThresh = 1e-05,
  ransacFinalSelectThresh = 2 * 1e-05,
  ransacIters = 150,
  useResiduals = TRUE,
  croppingThresh = 1,
  standardizeBF = FALSE,
  m = 2,
  mY = m,
  offset = 0,
  offsetY = offset,
  gaussFilterRes,
  gaussFilterWavelength = c(16, 250),
  gaussFilterType = "bp"
)
}
\arguments{
\item{x3p_path}{path to an .x3p file}

\item{ransacInlierThresh}{threshold to declare an observed value close to the
fitted plane an "inlier" for the RANSAC method}

\item{ransacFinalSelectThresh}{once the RANSAC plane is fitted based on the
inlierThreshold, this argument dictates which observations are selected as
the final breech face estimate.}

\item{ransacIters}{number of candidate planes to fit for the RANSAC method
(higher value yields more stable breech face estimate)}

\item{useResiduals}{dictates whether the difference between the estimated
breech face and fitted plane are returned (residuals) or if the estimates
breech face is simply shifted down by its mean value}

\item{croppingThresh}{treshold for cropping exterior whitespace from the
breech face impression}

\item{standardizeBF}{subtract mean and divide by standard deviation of
processed surface matrix}

\item{m}{integer value - every mth value is included in the sample}

\item{mY}{integer value - every mth value is included in the sample in the x
direction and every mYth value is included in y direction}

\item{offset}{integer value between 0 and m-1 to specify offset of the sample}

\item{offsetY}{integer value between 0 and mY-1 to specify different offsets
for x and y direction.}

\item{gaussFilterRes}{sampling resolution of the surface matrix
(given by incrementX or incrementY of the header.info element of an x3p
object). If not given, then this is determined automatically within the
function}

\item{gaussFilterWavelength}{cut-off wavelength(s) to be attenuated}

\item{gaussFilterType}{specifies whether a low pass, "lp", high
pass, "hp", or bandpass, "bp" filter is to be used. Note that setting
filterype = "bp" means that wavelength should be a vector of two numbers.
In this case, the max of these two number will be used for the high pass
filter and the min for the low pass filter.}
}
\value{
x3p object containing the processed, down-sampled breech face
}
\description{
Select breech face impression from a cartridge case scan after using
x3ptools::sample_x3p to downsample the scan.
}
\note{
Given a matrix, x3ptools populates an x3p object's surface matrix
starting in the top left corner moving right by reading the the matrix from
the bottom left corner and moving up. Effectively rotating the matrix by 90
degrees clockwise.
}
\examples{
\dontrun{
 #downsample x3p to a quarter the original size (every other row/col selected)
 processedBF1 <- cmcR::selectBFImpression_sample_x3p(x3p_path = "path/to/file.x3p",m = 2)
}
}
\seealso{
cartridges3D package (LINK)

x3ptools package (LINK)
}
