% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{x3pListPlot}
\alias{x3pListPlot}
\title{Plot a list of x3ps}
\arguments{
\item{x3pList}{a list of x3p objects. If the x3p objects are named in the
list, then these names will be included in the title of their respective
plot}

\item{type}{dictates whether one plot faceted by surface matrix or a list of
plots per surface matrix is returned. The faceted plot will have a
consistent height scale across all surface matrices.}

\item{rotate}{angle (in degrees) to rotate all surface matrices plotted}

\item{legend.quantiles}{vector of quantiles to be shown as tick marks on
legend plot}

\item{height.colors}{vector of colors to be passed to scale_fill_gradientn
that dictates the height value colorscale}

\item{na.value}{color to be used for NA values (passed to
scale_fill_gradientn)}

\item{guide}{internal usage}
}
\description{
Plots the surface matrices in a list of x3p objects. Either
creates one plot faceted by surface matrix or creates individual plots per
surface matrix and returns them in a list.
}
\examples{
data(fadul1.1_processed,fadul1.2_processed,fadul2.1_processed)

x3pListPlot(list("Fadul 1-1" = fadul1.1_processed,
                 "Fadul 1-2" = fadul1.2_processed,
                 "Fadul 2-1" = fadul2.1_processed))

}
