% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticTools.R
\name{x3pListPlot}
\alias{x3pListPlot}
\title{Plot a list of x3ps}
\arguments{
\item{x3pList}{a list of x3p objects. If the x3p objects are named in the
list, then these names will be included in the title of their respective
plot}

\item{type}{dictates whether one plot faceted by surface matrix or a list of
plots per surface matrix is returned. The faceted plot will have a
consistent height scale across all surface matrices.}

\item{rotate}{angle (in degrees) to rotate all surface matrices plotted}
}
\description{
Plots the surface matrices in a list of x3p objects. Either
creates one plot faceted by surface matrix or creates individual plots per
surface matrix and returns them in a list.
}
\examples{
\dontrun{
 x3pListPlot(list("name1" = x3p1, "name2" = x3p2))
}
}
